+++ apicast/src/resty/http_ng/response.lua
@@ -22,8 +22,8 @@ response.headers = require 'resty.http_ng.headers'
 
 
 function response.new(request, status, headers, body)
-  assert(status)
-  assert(body)
+  assert(status, 'missing request status')
+  assert(body, 'missing request body')
 
   local mt = {}
   mt['__index'] = function(table, key)
@@ -47,10 +47,19 @@ function response.new(request, status, headers, body)
     mt.body = body
   end
 
+  -- https://tools.ietf.org/html/rfc7231#section-7.1.1.2
+  -- > A recipient with a clock that receives a response message without a
+  --   Date header field MUST record the time it was received and append a
+  --   corresponding Date header field to the message's header section if it
+  --   is cached or forwarded downstream.
+  res.headers.date = res.headers.date or ngx.http_time(ngx.time())
+
   return setmetatable(res, mt)
 end
 
-function response.error(message, request)
+function response.error(request, message)
+  assert(request, 'missing request')
+  assert(message, 'missing message')
   return { ok = false, error = message, request = request, status = 0, headers = {} }
 end