+++ spec/keycloak_spec.lua
@@ -1,6 +1,7 @@
 local env = require 'resty.env'
 local _M = require 'oauth.keycloak'
 local test_backend_client = require 'resty.http_ng.backend.test'
+local jwt = require 'resty.jwt'
 
 describe('Keycloak', function()
     local test_backend
@@ -50,11 +51,11 @@ describe('Keycloak', function()
       stub(ngx.req, 'get_uri_args', function() return { response_type = 'code', client_id = 'foo', redirect_uri = 'bar' } end)
 
       test_backend.expect{ url = 'http://www.example.com:80/auth/realms/test/protocol/openid-connect/auth?client_id=foo' }
-        .respond_with{ status = 200 , body = 'foo', headers = {} }
+        .respond_with{ status = 200 , body = 'foo', headers = { Date = "Thu, 18 Nov 2010 11:27:35 GMT" } }
 
       stub(_M, 'respond_and_exit')
       keycloak:authorize({}, test_backend)
-      assert.spy(_M.respond_and_exit).was.called_with(200, 'foo', {})
+      assert.spy(_M.respond_and_exit).was.called_with(200, 'foo', { Date = "Thu, 18 Nov 2010 11:27:35 GMT" })
     end)
 
     it('returns error when response_type missing', function()
@@ -97,11 +98,11 @@ describe('Keycloak', function()
       stub(ngx.req, 'get_headers', function() return { } end)
 
       test_backend.expect{ url = 'http://www.example.com:80/auth/realms/test/protocol/openid-connect/token'}
-        .respond_with{ status = 200 , body = 'foo', headers = {} }
+        .respond_with{ status = 200 , body = 'foo', headers = { Date = "Thu, 18 Nov 2010 11:27:35 GMT" } }
 
       keycloak:get_token({}, test_backend)
 
-      assert.spy(_M.respond_and_exit).was.called_with(200, 'foo', {})
+      assert.spy(_M.respond_and_exit).was.called_with(200, 'foo', { Date = "Thu, 18 Nov 2010 11:27:35 GMT" })
     end)
 
     it('returns "invalid_request" when grant_type missing', function ()
@@ -170,11 +171,11 @@ describe('Keycloak', function()
       stub(ngx.req, 'get_headers', function() return { } end)
 
       test_backend.expect{ url = 'http://www.example.com:80/auth/realms/test/protocol/openid-connect/token'}
-        .respond_with{ status = 200 , body = 'foo', headers = {} }
+        .respond_with{ status = 200 , body = 'foo', headers = { Date = "Thu, 18 Nov 2010 11:27:35 GMT" } }
 
       keycloak:get_token({}, test_backend)
 
-      assert.spy(_M.respond_and_exit).was.called_with(200, 'foo', {})
+      assert.spy(_M.respond_and_exit).was.called_with(200, 'foo', { Date = "Thu, 18 Nov 2010 11:27:35 GMT" })
     end)
 
     it('accepts client credentials in Authorization header', function()
@@ -193,11 +194,11 @@ describe('Keycloak', function()
       test_backend.expect{
         url = 'http://www.example.com:80/auth/realms/test/protocol/openid-connect/token',
         headers =  { authorization = auth }
-      }.respond_with{ status = 200 , body = 'foo', headers = {} }
+      }.respond_with{ status = 200 , body = 'foo', headers = { Date = "Thu, 18 Nov 2010 11:27:35 GMT" } }
 
       keycloak:get_token({}, test_backend)
 
-      assert.spy(_M.respond_and_exit).was.called_with(200, 'foo', {})
+      assert.spy(_M.respond_and_exit).was.called_with(200, 'foo', { Date = "Thu, 18 Nov 2010 11:27:35 GMT" })
     end)
 
   end)
@@ -235,5 +236,34 @@ describe('Keycloak', function()
       assert.is.falsy(_M.token_get_headers()['foo'])
     end)
 
+  end)
+
+  describe('.transform_credentials', function()
+
+    it('returns ttl if exp', function()
+      local keycloak = _M.new(configuration)
+      local jwt_obj = { payload = { aud = "f8c6069e", exp = 1490705281 }, verified = true }
+
+      stub(jwt, 'verify', function() return jwt_obj end)
+      stub(ngx, 'now', function() return 1490705200 end)
+
+      local creds, ttl = keycloak:transform_credentials("foo")
+
+      assert.are.same({ app_id = 'f8c6069e' }, creds)
+      assert.equals(81, ttl )
+    end)
+
+     it('returns nil if no exp', function()
+      local keycloak = _M.new(configuration)
+      local jwt_obj = { payload = { aud = "f8c6069e" }, verified = true }
+
+      stub(jwt, 'verify', function() return jwt_obj end)
+      stub(ngx, 'now', function() return 1490705200 end)
+
+      local creds, ttl = keycloak:transform_credentials("foo")
+
+      assert.are.same({ app_id = 'f8c6069e' }, creds)
+      assert.equals(nil, ttl )
+    end)
   end)
 end)