+++ spec/resty/http_ng/backend/ngx_spec.lua
@@ -11,7 +11,7 @@ describe('ngx backend',function()
         assert.are.same({_url = 'http://localhost:8081'}, options.vars)
         return { status = 200, body = '' }
       end
-      local response = backend.send{method = method, url = 'http://localhost:8081' }
+      local response = backend:send{method = method, url = 'http://localhost:8081' }
       assert.truthy(response)
       assert.truthy(response.request)
     end)
@@ -23,7 +23,7 @@ describe('ngx backend',function()
         assert.are.same({Host = 'fake.example.com'}, options.ctx.headers)
         return { status = 200, body = '' }
       end
-      local response = backend.send{method = method, url = 'http://localhost:8081/path', headers = {Host = 'fake.example.com'} }
+      local response = backend:send{method = method, url = 'http://localhost:8081/path', headers = {Host = 'fake.example.com'} }
       assert.truthy(response)
       assert.truthy(response.request)
     end)