+++ spec/resty/http_ng/backend/resty_spec.lua
@@ -6,7 +6,7 @@ describe('resty backend', function()
     local method = 'GET'
 
     it('accesses the url', function()
-      local response, err = backend.send{method = method, url = 'http://example.com/'}
+      local response, err = backend:send{method = method, url = 'http://example.com/'}
       assert.falsy(err)
       assert.falsy(response.error)
       assert.truthy(response.ok)
@@ -14,7 +14,7 @@ describe('resty backend', function()
     end)
 
     it('works with ssl', function()
-      local response, err = backend.send{
+      local response, err = backend:send{
         method = method, url = 'https://google.com/',
         -- This is needed because of https://groups.google.com/forum/#!topic/openresty-en/SuqORBK9ys0
         -- So far OpenResty can't use system certificated on demand.
@@ -25,6 +25,16 @@ describe('resty backend', function()
       assert.truthy(response.ok)
       assert.truthy(response.request)
       assert(response.headers.location:match('^https://'))
+    end)
+
+    it('returns error', function()
+      local req = { method = method, url = 'http://0.0.0.0:0/' }
+      local response, err = backend:send(req)
+
+      assert.falsy(err)
+      assert.truthy(response.error)
+      assert.falsy(response.ok)
+      assert.same(req, response.request)
     end)
   end)
 end)